requirements.txt
Django==3.2
djangorestframework==3.12
cryptography==3.4
requests==2.26